SAMA-48 add attachments components
