+++ src/components/generic/messageComponents/MessageAttachment.js
@@ -1,7 +1,7 @@
-import { Blurhash } from "react-blurhash";
-import { Oval } from "react-loader-spinner";
+import ImageView from "../attachmentComponents/ImageView";
+import VideoView from "../attachmentComponents/VideoView";
 import { useLocation, useNavigate } from "react-router-dom";
-import { useMemo, useRef, useState } from "react";
+import { useRef } from "react";
 
 export default function MessageAttachment({
   id,
@@ -13,87 +13,30 @@ export default function MessageAttachment({
   const { hash } = useLocation();
   const navigate = useNavigate();
 
-  const [loaded, setLoaded] = useState(false);
   const videoRef = useRef(null);
-
-  const videoView = useMemo(() => {
-    const video = (
-      <video
-        ref={videoRef}
-        controls
-        src={url + "#t=0.1"}
-        poster={name}
-        onClick={(event) => {
-          event.preventDefault();
-          videoRef.current.pause();
-        }}
-      />
-    );
-
-    //get first clip from video for bg
-
-    return video;
-  }, [url]);
-
-  const preloaderView = useMemo(() => {
-    if (loaded || !blurHash) {
-      return null;
-    }
-
-    return localUrl ? (
-      <img src={localUrl} alt={name} />
-    ) : (
-      <div className="blur-hash-preloader">
-        <Blurhash
-          className="canvas-preloader"
-          key={id}
-          hash={blurHash}
-          width={400}
-          height={300}
-          resolutionX={32}
-          resolutionY={32}
-        />
-        <Oval
-          height={50}
-          width={50}
-          color="#1a8ee1"
-          wrapperClass={"blur-hash-loader"}
-          visible={true}
-          ariaLabel="oval-loading"
-          secondaryColor="#8dc7f0"
-          strokeWidth={2}
-          strokeWidthSecondary={3}
-        />
-      </div>
-    );
-  }, [loaded, localUrl, blurHash]);
 
   return (
     <div
       className="attachment-img"
+      style={{ gridColumnEnd: `span ${1}`, gridRowEnd: `span ${1}` }}
       onClick={() => {
         navigate(hash + `/modal?id=${id.replaceAll(" ", "%")}`);
-        if (videoRef.current) {
+        videoRef.current &&
           localStorage.setItem(
             "currentTimeVideo",
             videoRef.current.currentTime
           );
-        }
       }}
-      style={{ gridColumnEnd: `span ${1}`, gridRowEnd: `span ${1}` }}
     >
       {name.includes(".mp4") ? (
-        videoView
+        <VideoView vRef={videoRef} url={url} posterName={name} />
       ) : (
-        <>
-          <img
-            style={loaded ? {} : { display: "none" }}
-            onLoad={() => setLoaded(true)}
-            src={url}
-            alt={name}
-          />
-          {preloaderView}
-        </>
+        <ImageView
+          url={url}
+          localUrl={localUrl}
+          blurHash={blurHash}
+          altName={name}
+        />
       )}
     </div>
   );
